Refactor the ellipsis helper function out of wordOccurrenceHelpers.js to its own file/module
