+++ __tests__/wordOccurrenceHelpers.test.js
@@ -1,10 +1,7 @@
 jest.unmock('fs-extra')
 import path from 'path-extra'
 import ManageResource from '../src/ManageResource'
-import {
-  getWordOccurrencesForQuote,
-  getQuoteOmittedString
-} from '../src/wordOccurrenceHelpers'
+import { getWordOccurrencesForQuote } from '../src/helpers/wordOccurrenceHelpers'
 
 function getTestResult(bookId, chapter, verse, quote) {
   const originalBiblePath = path.join('__tests__', 'fixtures', 'resources', 'el-x-koine', 'bibles', 'ugnt', 'v0.5')
@@ -14,157 +11,6 @@ function getTestResult(bookId, chapter, verse, quote) {
 
   return words;
 }
-
-describe('getQuoteOmittedString():', () => {
-  test('it gets all the omitted words in a quote with one or more ellipsis', () => {
-    const checks = [
-      {// mat 1:7 figs-ellipsis
-        quote: 'Ῥοβοάμ, Ῥοβοὰμ ... ἐγέννησεν τὸν Ἀβιά, Ἀβιὰ ... ἐγέννησεν τὸν Ἀσάφ',
-        wholeQuote: "Ῥοβοάμ, Ῥοβοὰμ δὲ ἐγέννησεν τὸν Ἀβιά, Ἀβιὰ δὲ ἐγέννησεν τὸν Ἀσάφ",
-        verseString: "Σολομὼν δὲ ἐγέννησεν τὸν Ῥοβοάμ, Ῥοβοὰμ δὲ ἐγέννησεν τὸν Ἀβιά, Ἀβιὰ δὲ ἐγέννησεν τὸν Ἀσάφ,",
-      },
-      {// TIT 1:15 figs-metaphor
-        quote: "τοῖς ... μεμιαμμένοις καὶ ἀπίστοις, οὐδὲν καθαρόν",
-        wholeQuote: "τοῖς δὲ μεμιαμμένοις καὶ ἀπίστοις, οὐδὲν καθαρόν",
-        verseString: "πάντα καθαρὰ τοῖς καθαροῖς; τοῖς δὲ μεμιαμμένοις καὶ ἀπίστοις, οὐδὲν καθαρόν; ἀλλὰ μεμίανται αὐτῶν καὶ ὁ νοῦς, καὶ ἡ συνείδησις.",
-      },
-      {// MAT 3:2 figs-metonymy
-        quote: "ἤγγικεν ... ἡ ... τῶν Οὐρανῶν",
-        wholeQuote: "ἤγγικεν γὰρ ἡ Βασιλεία τῶν Οὐρανῶν",
-        verseString: "λέγων, \"μετανοεῖτε, ἤγγικεν γὰρ ἡ Βασιλεία τῶν Οὐρανῶν.\"",
-      },
-      {// MAT	12	19	hb2m	figs-metonymy	οὐδὲ ... ἀκούσει τις ... τὴν φωνὴν αὐτοῦ
-        quote: "οὐδὲ ... ἀκούσει τις ... τὴν φωνὴν αὐτοῦ",
-        wholeQuote: "οὐδὲ κραυγάσει, οὐδὲ ἀκούσει τις ἐν ταῖς πλατείαις τὴν φωνὴν αὐτοῦ",
-        verseString: "οὐκ ἐρίσει, οὐδὲ κραυγάσει, οὐδὲ ἀκούσει τις ἐν ταῖς πλατείαις τὴν φωνὴν αὐτοῦ.",
-      },
-      {// MAT	10	17	a55q	writing-connectingwords	προσέχετε ... ἀπὸ ... ἀνθρώπων; παραδώσουσιν
-        quote: "προσέχετε ... ἀπὸ ... ἀνθρώπων; παραδώσουσιν",
-        wholeQuote: "προσέχετε δὲ ἀπὸ τῶν ἀνθρώπων; παραδώσουσιν",
-        verseString: "προσέχετε δὲ ἀπὸ τῶν ἀνθρώπων; παραδώσουσιν γὰρ ὑμᾶς εἰς συνέδρια, καὶ ἐν ταῖς συναγωγαῖς αὐτῶν μαστιγώσουσιν ὑμᾶς;",
-      },
-      {// MAT	18	1	pp31	figs-metonymy	ἐν ... τῇ ... Βασιλεία τῶν Οὐρανῶν
-        quote: "ἐν ... τῇ ... Βασιλεία τῶν Οὐρανῶν",
-        wholeQuote: "ἐν ἐκείνῃ τῇ ὥρᾳ προσῆλθον οἱ μαθηταὶ τῷ Ἰησοῦ λέγοντες, \"τίς ἄρα μείζων ἐστὶν ἐν τῇ Βασιλεία τῶν Οὐρανῶν",
-        verseString: "ἐν ἐκείνῃ τῇ ὥρᾳ προσῆλθον οἱ μαθηταὶ τῷ Ἰησοῦ λέγοντες, \"τίς ἄρα μείζων ἐστὶν ἐν τῇ Βασιλεία τῶν Οὐρανῶν?\"",
-      },
-      {// MAT	19	12	m1r9	figs-activepassive	εἰσὶν ... εὐνοῦχοι, οἵτινες ... εὐνουχίσθησαν ὑπὸ τῶν ἀνθρώπων
-        quote: "εἰσὶν ... εὐνοῦχοι, οἵτινες ... εὐνουχίσθησαν ὑπὸ τῶν ἀνθρώπων",
-        wholeQuote: "εἰσὶν γὰρ εὐνοῦχοι, οἵτινες ἐκ κοιλίας μητρὸς ἐγεννήθησαν οὕτως, καὶ εἰσὶν εὐνοῦχοι οἵτινες εὐνουχίσθησαν ὑπὸ τῶν ἀνθρώπων",
-        verseString: "εἰσὶν γὰρ εὐνοῦχοι, οἵτινες ἐκ κοιλίας μητρὸς ἐγεννήθησαν οὕτως, καὶ εἰσὶν εὐνοῦχοι οἵτινες εὐνουχίσθησαν ὑπὸ τῶν ἀνθρώπων, καὶ εἰσὶν εὐνοῦχοι οἵτινες εὐνούχισαν ἑαυτοὺς διὰ τὴν Βασιλείαν τῶν Οὐρανῶν. ὁ δυνάμενος χωρεῖν, χωρείτω.\"",
-      },
-      {// MAT	21	9	g73z	figs-metonymy	ὡσαννὰ ... ἐν ... τοῖς ὑψίστοις
-        quote: "ὡσαννὰ ... ἐν ... τοῖς ὑψίστοις",
-        wholeQuote: "ὡσαννὰ τῷ Υἱῷ Δαυείδ! εὐλογημένος‘ ὁ ἐρχόμενος ἐν ὀνόματι Κυρίου!’ ὡσαννὰ ἐν τοῖς ὑψίστοις",
-        verseString: "οἱ δὲ ὄχλοι οἱ προάγοντες αὐτὸν καὶ οἱ ἀκολουθοῦντες ἔκραζον λέγοντες, \"ὡσαννὰ τῷ Υἱῷ Δαυείδ! εὐλογημένος‘ ὁ ἐρχόμενος ἐν ὀνόματι Κυρίου!’ ὡσαννὰ ἐν τοῖς ὑψίστοις!\"",
-      },
-      {// MAT	5	29	et3n	figs-you	εἰ ... ὁ ὀφθαλμός σου ὁ δεξιὸς
-        quote: "εἰ ... ὁ ὀφθαλμός σου ὁ δεξιὸς",
-        wholeQuote: "εἰ δὲ ὁ ὀφθαλμός σου ὁ δεξιὸς",
-        verseString: "εἰ δὲ ὁ ὀφθαλμός σου ὁ δεξιὸς σκανδαλίζει σε, ἔξελε αὐτὸν καὶ βάλε ἀπὸ σοῦ; συμφέρει γάρ σοι ἵνα ἀπόληται ἓν τῶν μελῶν σου, καὶ μὴ ὅλον τὸ σῶμά σου βληθῇ εἰς Γέενναν.",
-      },
-      {// MAT	2	13 figs-explicit ἕως ... εἴπω σοι
-        quote: "ἕως ... εἴπω σοι",
-        wholeQuote: "ἕως ἂν εἴπω σοι",
-        verseString: "ἀναχωρησάντων δὲ αὐτῶν, ἰδοὺ, ἄγγελος Κυρίου φαίνεται κατ’ ὄναρ τῷ Ἰωσὴφ λέγων, \"ἐγερθεὶς, παράλαβε τὸ παιδίον καὶ τὴν μητέρα αὐτοῦ καὶ φεῦγε εἰς Αἴγυπτον, καὶ ἴσθι ἐκεῖ ἕως ἂν εἴπω σοι; μέλλει γὰρ Ἡρῴδης ζητεῖν τὸ παιδίον τοῦ ἀπολέσαι αὐτό.\"",
-      },
-      {// COL	1	28	va1x	figs-exclusive	ἡμεῖς καταγγέλλομεν, νουθετοῦντες ... διδάσκοντες ... παραστήσωμεν
-        quote: "ἡμεῖς καταγγέλλομεν, νουθετοῦντες ... διδάσκοντες ... παραστήσωμεν",
-        wholeQuote: "ἡμεῖς καταγγέλλομεν, νουθετοῦντες πάντα ἄνθρωπον καὶ διδάσκοντες πάντα ἄνθρωπον ἐν πάσῃ σοφίᾳ, ἵνα παραστήσωμεν",
-        verseString: "ὃν ἡμεῖς καταγγέλλομεν, νουθετοῦντες πάντα ἄνθρωπον καὶ διδάσκοντες πάντα ἄνθρωπον ἐν πάσῃ σοφίᾳ, ἵνα παραστήσωμεν πάντα ἄνθρωπον τέλειον ἐν Χριστῷ.",
-      },
-      {// ROM	1	11	f3g1	figs-explicit	τι ... χάρισμα ... πνευματικὸν, εἰς τὸ στηριχθῆναι ὑμᾶς
-        quote: "τι ... χάρισμα ... πνευματικὸν, εἰς τὸ στηριχθῆναι ὑμᾶς",
-        wholeQuote: "τι μεταδῶ χάρισμα ὑμῖν πνευματικὸν, εἰς τὸ στηριχθῆναι ὑμᾶς",
-        verseString: "ἐπιποθῶ γὰρ ἰδεῖν ὑμᾶς, ἵνα τι μεταδῶ χάρισμα ὑμῖν πνευματικὸν, εἰς τὸ στηριχθῆναι ὑμᾶς;",
-      },
-      {// ROM	1	17	h38h	figs-activepassive	δικαιοσύνη ... Θεοῦ ... ἀποκαλύπτεται, ἐκ πίστεως εἰς πίστιν
-        quote: "δικαιοσύνη ... Θεοῦ ... ἀποκαλύπτεται, ἐκ πίστεως εἰς πίστιν",
-        wholeQuote: "δικαιοσύνη γὰρ Θεοῦ ἐν αὐτῷ ἀποκαλύπτεται, ἐκ πίστεως εἰς πίστιν",
-        verseString: "δικαιοσύνη γὰρ Θεοῦ ἐν αὐτῷ ἀποκαλύπτεται, ἐκ πίστεως εἰς πίστιν; καθὼς γέγραπται, \"ὁ δὲ δίκαιος ἐκ πίστεως ζήσεται.\"",
-      },
-      {// ROM	2	24	q13d	figs-activepassive	τὸ ... ὄνομα τοῦ Θεοῦ ... βλασφημεῖται ἐν τοῖς ἔθνεσιν
-        quote: "τὸ ... ὄνομα τοῦ Θεοῦ ... βλασφημεῖται ἐν τοῖς ἔθνεσιν",
-        wholeQuote: "τὸ γὰρ ὄνομα τοῦ Θεοῦ δι’ ὑμᾶς βλασφημεῖται ἐν τοῖς ἔθνεσιν",
-        verseString: "\"τὸ γὰρ ὄνομα τοῦ Θεοῦ δι’ ὑμᾶς βλασφημεῖται ἐν τοῖς ἔθνεσιν\", καθὼς γέγραπται.",
-      },
-      {// ROM	4	12	s9jt	figs-idiom	τοῖς ... στοιχοῦσιν τοῖς ἴχνεσιν τῆς ... πίστεως, τοῦ πατρὸς ἡμῶν Ἀβραάμ
-        quote: "τοῖς ... στοιχοῦσιν τοῖς ἴχνεσιν τῆς ... πίστεως, τοῦ πατρὸς ἡμῶν Ἀβραάμ",
-        wholeQuote: "τοῖς οὐκ ἐκ περιτομῆς μόνον, ἀλλὰ καὶ τοῖς στοιχοῦσιν τοῖς ἴχνεσιν τῆς ἐν ἀκροβυστίᾳ πίστεως, τοῦ πατρὸς ἡμῶν Ἀβραάμ",
-        verseString: "καὶ πατέρα περιτομῆς τοῖς οὐκ ἐκ περιτομῆς μόνον, ἀλλὰ καὶ τοῖς στοιχοῦσιν τοῖς ἴχνεσιν τῆς ἐν ἀκροβυστίᾳ πίστεως, τοῦ πατρὸς ἡμῶν Ἀβραάμ.",
-      },
-      {// ROM	7	2	l6d9	figs-metaphor	ἡ ... ὕπανδρος γυνὴ τῷ ... ἀνδρὶ δέδεται νόμῳ
-        quote: "ἡ ... ὕπανδρος γυνὴ τῷ ... ἀνδρὶ δέδεται νόμῳ",
-        wholeQuote: "ἡ γὰρ ὕπανδρος γυνὴ τῷ ζῶντι ἀνδρὶ δέδεται νόμῳ",
-        verseString: " ἡ γὰρ ὕπανδρος γυνὴ τῷ ζῶντι ἀνδρὶ δέδεται νόμῳ; ἐὰν δὲ ἀποθάνῃ ὁ ἀνήρ, κατήργηται ἀπὸ τοῦ νόμου τοῦ ἀνδρός.",
-      },
-      {// ROM	8	19	d911	figs-personification	ἡ ... ἀποκαραδοκία τῆς κτίσεως, τὴν ... ἀπεκδέχεται
-        quote: "ἡ ... ἀποκαραδοκία τῆς κτίσεως, τὴν ... ἀπεκδέχεται",
-        wholeQuote: "ἡ γὰρ ἀποκαραδοκία τῆς κτίσεως, τὴν ἀποκάλυψιν τῶν υἱῶν τοῦ Θεοῦ ἀπεκδέχεται",
-        verseString: "ἡ γὰρ ἀποκαραδοκία τῆς κτίσεως, τὴν ἀποκάλυψιν τῶν υἱῶν τοῦ Θεοῦ ἀπεκδέχεται.",
-      },
-      {// ROM	14	7	c9ls	figs-explicit	οὐδεὶς ... ἑαυτῷ ... ἀποθνῄσκει
-        quote: "οὐδεὶς ... ἑαυτῷ ... ἀποθνῄσκει",
-        wholeQuote: "οὐδεὶς γὰρ ἡμῶν ἑαυτῷ ζῇ, καὶ οὐδεὶς ἑαυτῷ ἀποθνῄσκει",
-        verseString: "οὐδεὶς γὰρ ἡμῶν ἑαυτῷ ζῇ, καὶ οὐδεὶς ἑαυτῷ ἀποθνῄσκει.",
-      },
-      {// REV	2	18	zbx5	figs-simile	ὁ ... ἔχων τοὺς ὀφθαλμοὺς ... ὡς φλόγα πυρός
-        quote: "ὁ ... ἔχων τοὺς ὀφθαλμοὺς ... ὡς φλόγα πυρός",
-        wholeQuote: "ὁ Υἱὸς τοῦ Θεοῦ, ὁ ἔχων τοὺς ὀφθαλμοὺς αὐτοῦ ὡς φλόγα πυρός",
-        verseString: "καὶ τῷ ἀγγέλῳ τῆς ἐν Θυατείροις ἐκκλησίας γράψον: τάδε‘ λέγει ὁ Υἱὸς τοῦ Θεοῦ, ὁ ἔχων τοὺς ὀφθαλμοὺς αὐτοῦ ὡς φλόγα πυρός, καὶ οἱ πόδες αὐτοῦ ὅμοιοι χαλκολιβάνῳ:",
-      },
-      {// MAT	28	7	sp2a	figs-quotesinquotes	εἴπατε τοῖς μαθηταῖς αὐτοῦ ... ἠγέρθη ἀπὸ τῶν νεκρῶν ... ἰδοὺ, προάγει ὑμᾶς εἰς τὴν Γαλιλαίαν; ἐκεῖ αὐτὸν ὄψεσθε
-        quote: "εἴπατε τοῖς μαθηταῖς αὐτοῦ ... ἠγέρθη ἀπὸ τῶν νεκρῶν ... ἰδοὺ, προάγει ὑμᾶς εἰς τὴν Γαλιλαίαν; ἐκεῖ αὐτὸν ὄψεσθε",
-        wholeQuote: "εἴπατε τοῖς μαθηταῖς αὐτοῦ, ὅτι ἠγέρθη ἀπὸ τῶν νεκρῶν; καὶ ἰδοὺ, προάγει ὑμᾶς εἰς τὴν Γαλιλαίαν; ἐκεῖ αὐτὸν ὄψεσθε",
-        verseString: "καὶ ταχὺ πορευθεῖσαι, εἴπατε τοῖς μαθηταῖς αὐτοῦ, ὅτι ἠγέρθη ἀπὸ τῶν νεκρῶν; καὶ ἰδοὺ, προάγει ὑμᾶς εἰς τὴν Γαλιλαίαν; ἐκεῖ αὐτὸν ὄψεσθε. ἰδοὺ, εἶπον ὑμῖν.\""
-      },
-       {// COL	1	9	f2xd	figs-exclusive	ἡμεῖς ... ἠκούσαμεν ... καὶ αἰτούμενοι
-        quote: "ἡμεῖς ... ἠκούσαμεν ... καὶ αἰτούμενοι",
-        wholeQuote: "ἡμεῖς, ἀφ’ ἧς ἡμέρας ἠκούσαμεν, οὐ παυόμεθα ὑπὲρ ὑμῶν προσευχόμενοι καὶ αἰτούμενοι",
-        verseString: "διὰ τοῦτο καὶ ἡμεῖς, ἀφ’ ἧς ἡμέρας ἠκούσαμεν, οὐ παυόμεθα ὑπὲρ ὑμῶν προσευχόμενοι καὶ αἰτούμενοι, ἵνα πληρωθῆτε τὴν ἐπίγνωσιν τοῦ θελήματος αὐτοῦ ἐν πάσῃ σοφίᾳ καὶ συνέσει πνευματικῇ,",
-      },
-      {// MAT 3:5 figs-metonymy
-        quote: "τότε ... Ἱεροσόλυμα ... πᾶσα ἡ Ἰουδαία, καὶ πᾶσα ἡ περίχωρος",
-        wholeQuote: "τότε ἐξεπορεύετο πρὸς αὐτὸν Ἱεροσόλυμα, καὶ πᾶσα ἡ Ἰουδαία, καὶ πᾶσα ἡ περίχωρος",
-        verseString: "τότε ἐξεπορεύετο πρὸς αὐτὸν Ἱεροσόλυμα, καὶ πᾶσα ἡ Ἰουδαία, καὶ πᾶσα ἡ περίχωρος τοῦ Ἰορδάνου,"
-      },
-      {// REV	9	15	b3d6	figs-parallelism	εἰς τὴν ὥραν ... ἡμέραν ... μῆνα, καὶ ἐνιαυτόν
-        quote: "εἰς τὴν ὥραν ... ἡμέραν ... μῆνα, καὶ ἐνιαυτόν",
-        wholeQuote: "εἰς τὴν ὥραν, καὶ ἡμέραν, καὶ μῆνα, καὶ ἐνιαυτόν",
-        verseString: "καὶ ἐλύθησαν οἱ τέσσαρες ἄγγελοι, οἱ ἡτοιμασμένοι εἰς τὴν ὥραν, καὶ ἡμέραν, καὶ μῆνα, καὶ ἐνιαυτόν, ἵνα ἀποκτείνωσιν τὸ τρίτον τῶν ἀνθρώπων.",
-      },
-      {// REV	12	9	pk5u	figs-distinguish	δράκων ὁ ... ὄφις ὁ ἀρχαῖος ... καλούμενος, Διάβολος, καὶ ὁ Σατανᾶς, ὁ πλανῶν τὴν οἰκουμένην ὅλην; ἐβλήθη εἰς τὴν γῆν, καὶ οἱ ἄγγελοι αὐτοῦ μετ’ αὐτοῦ ἐβλήθησαν
-        quote: "δράκων ὁ ... ὄφις ὁ ἀρχαῖος ... καλούμενος, Διάβολος, καὶ ὁ Σατανᾶς, ὁ πλανῶν τὴν οἰκουμένην ὅλην; ἐβλήθη εἰς τὴν γῆν, καὶ οἱ ἄγγελοι αὐτοῦ μετ’ αὐτοῦ ἐβλήθησαν",
-        wholeQuote: "δράκων ὁ μέγας, ὁ ὄφις ὁ ἀρχαῖος, ὁ καλούμενος, Διάβολος, καὶ ὁ Σατανᾶς, ὁ πλανῶν τὴν οἰκουμένην ὅλην; ἐβλήθη εἰς τὴν γῆν, καὶ οἱ ἄγγελοι αὐτοῦ μετ’ αὐτοῦ ἐβλήθησαν",
-        verseString: "καὶ ἐβλήθη ὁ δράκων ὁ μέγας, ὁ ὄφις ὁ ἀρχαῖος, ὁ καλούμενος, Διάβολος, καὶ ὁ Σατανᾶς, ὁ πλανῶν τὴν οἰκουμένην ὅλην; ἐβλήθη εἰς τὴν γῆν, καὶ οἱ ἄγγελοι αὐτοῦ μετ’ αὐτοῦ ἐβλήθησαν.",
-      },
-      {// MAT	3	13	zbj9	figs-activepassive	τὸν ... Ἰωάννην ... βαπτισθῆναι ὑπ’
-        quote: "τὸν ... Ἰωάννην ... βαπτισθῆναι ὑπ’",
-        wholeQuote: "τὸν Ἰορδάνην πρὸς τὸν Ἰωάννην, τοῦ βαπτισθῆναι ὑπ’",
-        verseString: "τότε παραγίνεται ὁ Ἰησοῦς ἀπὸ τῆς Γαλιλαίας ἐπὶ τὸν Ἰορδάνην πρὸς τὸν Ἰωάννην, τοῦ βαπτισθῆναι ὑπ’ αὐτοῦ."
-      },
-      {// REV	3	17	v1pj	figs-metaphor	σὺ εἶ ὁ ταλαίπωρος ... ἐλεεινὸς ... πτωχὸς ... τυφλὸς, καὶ γυμνός
-        quote: "σὺ εἶ ὁ ταλαίπωρος ... ἐλεεινὸς ... πτωχὸς ... τυφλὸς, καὶ γυμνός",
-        wholeQuote: "σὺ εἶ ὁ ταλαίπωρος, καὶ ἐλεεινὸς, καὶ πτωχὸς, καὶ τυφλὸς, καὶ γυμνός",
-        verseString: "ὅτι λέγεις, ὅτι πλούσιός‘ εἰμι, καὶ πεπλούτηκα, καὶ οὐδὲν χρείαν ἔχω’, καὶ οὐκ οἶδας ὅτι σὺ εἶ ὁ ταλαίπωρος, καὶ ἐλεεινὸς, καὶ πτωχὸς, καὶ τυφλὸς, καὶ γυμνός.",
-      },
-      {// COL	2	7	e2x6	figs-idiom	ἐρριζωμένοι ... ἐποικοδομούμενοι ... βεβαιούμενοι ... περισσεύοντες
-        quote: "ἐρριζωμένοι ... ἐποικοδομούμενοι ... βεβαιούμενοι ... περισσεύοντες",
-        wholeQuote: "ἐρριζωμένοι καὶ ἐποικοδομούμενοι ἐν αὐτῷ, καὶ βεβαιούμενοι τῇ πίστει, καθὼς ἐδιδάχθητε, περισσεύοντες",
-        verseString: "ἐρριζωμένοι καὶ ἐποικοδομούμενοι ἐν αὐτῷ, καὶ βεβαιούμενοι τῇ πίστει, καθὼς ἐδιδάχθητε, περισσεύοντες ἐν εὐχαριστίᾳ.",
-      },
-      {// MAT	3:5 figs-metonymy	τότε ... Ἱεροσόλυμα ... πᾶσα ἡ Ἰουδαία, καὶ πᾶσα ἡ περίχωρος
-        quote: "τότε ... Ἱεροσόλυμα ... πᾶσα ἡ Ἰουδαία, καὶ πᾶσα ἡ περίχωρος",
-        wholeQuote: "τότε ἐξεπορεύετο πρὸς αὐτὸν Ἱεροσόλυμα, καὶ πᾶσα ἡ Ἰουδαία, καὶ πᾶσα ἡ περίχωρος",
-        verseString: "τότε ἐξεπορεύετο πρὸς αὐτὸν Ἱεροσόλυμα, καὶ πᾶσα ἡ Ἰουδαία, καὶ πᾶσα ἡ περίχωρος τοῦ Ἰορδάνου,",
-      },
-    ]
-
-    checks.forEach(({ quote, verseString, wholeQuote}) => {
-      expect(getQuoteOmittedString(quote, verseString)).toEqual(wholeQuote)
-    })
-  })
-})
 
 describe('getWordOccurrencesForQuote():', () => {
   // test('', () => {
