+++ src/utils/constants.js
@@ -1,0 +1,2 @@
+export const ELLIPSES = '\u2026'
+export const THREE_DOTS = '...'
